Refactoring to make Atomix OSGi compliant
