+++ all/src/main/java/io/atomix/bundle/AtomixBundle.java
@@ -1,0 +1,10 @@
+package io.atomix.bundle;
+
+/**
+ * Empty class required to get the atomix-all module to build properly.
+ *
+ * NOTE Required for bundle plugin to operate.
+ */
+public class AtomixBundle {
+
+}
