+++ all/src/main/java/io/atomix/bundle/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Empty package for bundle plugin
+ */
+package io.atomix.bundle;
