+++ collections/src/main/java/io/atomix/collections/internal/MultiMapState.java
@@ -15,7 +15,7 @@
  */
 package io.atomix.collections.internal;
 
-import io.atomix.catalyst.util.concurrent.Scheduled;
+import io.atomix.catalyst.concurrent.Scheduled;
 import io.atomix.collections.DistributedMultiMap;
 import io.atomix.copycat.server.Commit;
 import io.atomix.resource.ResourceStateMachine;
