+++ examples/distributed-value/src/main/java/io/atomix/examples/variables/DistributedValueExample.java
@@ -17,7 +17,7 @@ package io.atomix.examples.variables;
 
 import io.atomix.AtomixReplica;
 import io.atomix.catalyst.transport.Address;
-import io.atomix.catalyst.transport.NettyTransport;
+import io.atomix.catalyst.transport.netty.NettyTransport;
 import io.atomix.copycat.server.storage.Storage;
 import io.atomix.copycat.server.storage.StorageLevel;
 import io.atomix.variables.DistributedValue;
