+++ groups/src/main/java/io/atomix/group/messaging/MessageConsumer.java
@@ -15,7 +15,7 @@
  */
 package io.atomix.group.messaging;
 
-import io.atomix.catalyst.util.Listener;
+import io.atomix.catalyst.concurrent.Listener;
 
 import java.util.function.Consumer;
