+++ manager/src/main/java/io/atomix/manager/internal/ResourceManagerState.java
@@ -23,10 +23,11 @@ import io.atomix.copycat.server.session.ServerSession;
 import io.atomix.copycat.server.session.SessionListener;
 import io.atomix.copycat.server.storage.snapshot.SnapshotReader;
 import io.atomix.copycat.server.storage.snapshot.SnapshotWriter;
+import io.atomix.manager.ResourceManagerException;
+import io.atomix.manager.resource.internal.InstanceOperation;
 import io.atomix.resource.Resource;
 import io.atomix.resource.ResourceStateMachine;
 import io.atomix.resource.ResourceType;
-import io.atomix.resource.internal.InstanceOperation;
 
 import java.util.*;
 import java.util.function.Function;
