+++ manager/src/main/java/io/atomix/manager/resource/internal/InstanceTypeResolver.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 the original author or authors.
+ * Copyright 2016 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -11,30 +11,25 @@
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
- * limitations under the License.
+ * limitations under the License
  */
-package io.atomix;
+package io.atomix.manager.resource.internal;
+
+import io.atomix.catalyst.serializer.SerializableTypeResolver;
+import io.atomix.catalyst.serializer.SerializerRegistry;
 
 /**
- * Atomix exception.
+ * Resource instance serializable type resolver.
  *
  * @author <a href="http://github.com/kuujo>Jordan Halterman</a>
  */
-public class AtomixException extends RuntimeException {
+public class InstanceTypeResolver implements SerializableTypeResolver {
 
-  public AtomixException() {
-  }
-
-  public AtomixException(String message) {
-    super(message);
-  }
-
-  public AtomixException(String message, Throwable cause) {
-    super(message, cause);
-  }
-
-  public AtomixException(Throwable cause) {
-    super(cause);
+  @Override
+  public void resolve(SerializerRegistry registry) {
+    registry.register(InstanceCommand.class, -55);
+    registry.register(InstanceQuery.class, -56);
+    registry.register(InstanceEvent.class, -57);
   }
 
 }
