+++ manager/src/main/java/io/atomix/manager/resource/internal/ResourceInstance.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License
  */
-package io.atomix.resource.internal;
+package io.atomix.manager.resource.internal;
 
 import io.atomix.catalyst.util.Assert;
 import io.atomix.resource.ResourceType;
