+++ manager/src/test/java/io/atomix/manager/options/ServerOptionsTest.java
@@ -15,7 +15,7 @@
  */
 package io.atomix.manager.options;
 
-import io.atomix.catalyst.transport.NettyTransport;
+import io.atomix.catalyst.transport.netty.NettyTransport;
 import io.atomix.catalyst.transport.Transport;
 import io.atomix.catalyst.util.PropertiesReader;
 import io.atomix.copycat.client.CopycatClient;
