+++ pom.xml
@@ -22,8 +22,8 @@
     <java.version>1.8</java.version>
     <slf4j.version>1.7.7</slf4j.version>
     <logback.version>1.1.2</logback.version>
-    <catalyst.version>1.0.8</catalyst.version>
-    <copycat.version>1.0.1</copycat.version>
+    <catalyst.version>1.0.9-SNAPSHOT</catalyst.version>
+    <copycat.version>1.0.2-SNAPSHOT</copycat.version>
 
     <maven.source.plugin.version>2.2.1</maven.source.plugin.version>
     <maven.compiler.plugin.version>3.0</maven.compiler.plugin.version>
