+++ resource/src/main/java/io/atomix/resource/Resource.java
@@ -15,10 +15,10 @@
  */
 package io.atomix.resource;
 
+import io.atomix.catalyst.concurrent.Listener;
+import io.atomix.catalyst.concurrent.ThreadContext;
 import io.atomix.catalyst.serializer.Serializer;
-import io.atomix.catalyst.util.Listener;
 import io.atomix.catalyst.util.Managed;
-import io.atomix.catalyst.util.concurrent.ThreadContext;
 import io.atomix.copycat.Command;
 import io.atomix.copycat.Query;
 import io.atomix.copycat.session.Session;
