+++ variables/src/main/java/io/atomix/variables/internal/ValueState.java
@@ -15,7 +15,7 @@
  */
 package io.atomix.variables.internal;
 
-import io.atomix.catalyst.util.concurrent.Scheduled;
+import io.atomix.catalyst.concurrent.Scheduled;
 import io.atomix.copycat.server.Commit;
 
 import java.time.Duration;
